message_delete: Refactor message delete functions in a new file.
